refactor: move methods to lib/*
BREAKING CHANGE:

separate methods like `require(pouchdb-hoodie-api)` can no longer be required directly
