+++ lib/add.js
