+++ lib/clear.js
