+++ lib/find-or-add.js
