+++ lib/find.js
