+++ lib/helpers/add-many.js
