+++ lib/helpers/add-one.js
