+++ lib/helpers/find-many.js
