+++ lib/helpers/find-one.js
