+++ lib/helpers/find-or-add-many.js
