+++ lib/helpers/update-many.js
