+++ lib/remove-all.js
