+++ lib/remove.js
