+++ lib/update-all.js
