+++ lib/utils/change-object.js
