+++ lib/utils/isnt-design-doc.js
