+++ lib/utils/mark-as-deleted.js
