+++ lib/utils/now.js
