+++ lib/utils/to-doc.js
