+++ lib/utils/to-object.js
