+++ lib/with-id-prefix.js
@@ -28,9 +28,9 @@ function withIdPrefix (state, prefix) {
     withIdPrefix: function (moarPrefix) {
       return withIdPrefix.call(db, state, prefix + moarPrefix)
     },
-    on: require('./lib/on').bind(this, {emitter: emitter}),
-    one: require('./lib/one').bind(this, {emitter: emitter}),
-    off: require('./lib/off').bind(this, {emitter: emitter})
+    on: require('./on').bind(this, {emitter: emitter}),
+    one: require('./one').bind(this, {emitter: emitter}),
+    off: require('./off').bind(this, {emitter: emitter})
   }
 
   state.emitter.on('change', function (eventName, object) {
