+++ tests/specs/remove-all.js
@@ -192,7 +192,7 @@ test('store.removeAll([objects]) creates deletedAt timestamps', function (t) {
   var db = dbFactory()
   var store = db.hoodieApi()
 
-  var now = require('../../utils/now')
+  var now = require('../../lib/utils/now')
   var isValidDate = require('../utils/is-valid-date')
 
   store.add([{
