+++ tests/specs/update-all.js
@@ -167,7 +167,7 @@ test('store.updateAll([objects]) updates all updatedAt timestamps', function (t)
   var db = dbFactory()
   var store = db.hoodieApi()
 
-  var now = require('../../utils/now')
+  var now = require('../../lib/utils/now')
   var isValidDate = require('../utils/is-valid-date')
 
   var updatedCount = 0
